pyproject.toml
[tool.poetry]
name = "Python FastAPI template"
version = "2.1.0"
description = "My personal best practice of archtecture in fastapi projects"
authors = ["AlanLatte <alan.latte.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
pydantic = {extras = ["dotenv", "email"], version = "^1.10.13"}
dependency-injector = "^4.40.0"
yoyo-migrations = "^8.1.0"
aiocache = {version = "^0.12.1", extras = ["redis"]}
aiopg = "^1.3.3"
bcrypt = "^4.0.1"
python-jose = {version = "3.3.0", extras = ["cryptography"]}
starlette-prometheus = "^0.9.0"
pydantic-yaml = "^0.11.2"
pymongo = "^4.2.0"
aiohttp = "^3.8.2"
pyotp = "^2.7.0"
httpx = "^0.24.1"
httpcore = "^0.16.2"
opentelemetry-instrumentation-fastapi = "^0.36b0"
opentelemetry-instrumentation-logging = "^0.36b0"
opentelemetry-exporter-otlp = "^1.15.0"
opentelemetry-api = "^1.15.0"
opentelemetry-sdk = "^1.15.0"
pyyaml = "^6.0"
jsf = "^0.7.1"

[tool.poetry.group.dev.dependencies]
MarkupSafe = "^2.1.0"

[tool.poetry.group.formatters.dependencies]
autoflake = "^1.4"
isort = "^5.10.1"
black = "^22.1.0"
docformatter = "^1.4"
add-trailing-comma = "^2.2.1"

[tool.poetry.group.sasts.dependencies]
itsdangerous = "^2.1.2"
safety = "^2.3.3"
bandit = "^1.7.1"

[tool.poetry.group.linters.dependencies]
flake8 = "^4.0.1"
mypy = "^0.931"
pylint = "^2.17.6"
ruff = "^0.1.8"
pylint-pytest = "^1.1.7"

[tool.poetry.group.tests.dependencies]
pytest-asyncio = "^0.20.2"
coverage = "^6.3.2"
pytest-cov = "^4.0.0"
pytest-repeat = "^0.9.1"

[tool.poetry.group.docs.dependencies]
Sphinx = "^5.3.0"
sphinx-pydantic = "^0.1.1"
sphinx-rtd-theme = "^1.1.1"


[tool.poetry.group.benchmarks.dependencies]
locust = "^2.14.2"


[tool.poetry.group.sugar.dependencies]
pytest-sugar = "^0.9.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
